+++ src/main/java/org/openset/automator/app/App.java
@@ -1,0 +1,4 @@
+package org.openset.automator.app;
+
+public interface App {
+}