+++ src/main/java/org/openset/automator/app/mobile/MobileApp.java
@@ -1,0 +1,6 @@
+package org.openset.automator.app.mobile;
+
+import org.openset.automator.app.App;
+
+public class MobileApp implements App {
+}