+++ src/main/java/org/openset/automator/app/mobile/MobilePage.java
@@ -1,0 +1,6 @@
+package org.openset.automator.app.mobile;
+
+import org.openset.automator.app.Page;
+
+public class MobilePage implements Page {
+}