+++ src/main/java/org/openset/automator/app/web/WebPage.java
@@ -1,0 +1,6 @@
+package org.openset.automator.app.web;
+
+import org.openset.automator.app.Page;
+
+public class WebPage implements Page {
+}