+++ src/main/java/org/openset/automator/app/web/WebsiteUnderTest.java
@@ -1,0 +1,6 @@
+package org.openset.automator.app.web;
+
+import org.openset.automator.app.App;
+
+public class WebsiteUnderTest implements App {
+}