+++ src/main/java/org/openset/automator/os/ExtractException.java
@@ -1,0 +1,9 @@
+package org.openset.automator.os;
+
+import java.io.IOException;
+
+public class ExtractException extends RuntimeException {
+    public ExtractException(String message, IOException exception) {
+        super(message, exception);
+    }
+}