+++ src/main/java/org/openset/automator/settings/SettingsLoadException.java
@@ -1,0 +1,9 @@
+package org.openset.automator.settings;
+
+import java.io.IOException;
+
+public class SettingsLoadException extends RuntimeException {
+    public SettingsLoadException(String message, IOException exception) {
+        super(message, exception);
+    }
+}