+++ src/main/java/org/openset/automator/settings/electron/ElectronSettings.java
@@ -1,0 +1,16 @@
+package org.openset.automator.settings.electron;
+
+import org.openset.automator.settings.BaseSettings;
+
+/**
+ * Settings in context of testing electron application.
+ */
+public class ElectronSettings {
+    public BaseSettings base;
+    public ElectronConfig electron;
+
+    public ElectronSettings() {
+        base = new BaseSettings();
+        electron = new ElectronConfig(base);
+    }
+}