+++ src/main/java/org/openset/automator/settings/web/WebSettings.java
@@ -1,0 +1,21 @@
+package org.openset.automator.settings.web;
+
+import org.openset.automator.settings.BaseSettings;
+import org.openset.automator.settings.cloud.SauceConfig;
+
+import java.io.IOException;
+
+/**
+ * Settings in context of org.automator.web testing.
+ */
+public class WebSettings {
+    public BaseSettings base;
+    public WebConfig web;
+    public SauceConfig sauce;
+
+    public WebSettings() throws IOException {
+        base = new BaseSettings();
+        web = new WebConfig(base);
+        sauce = new SauceConfig(base);
+    }
+}