+++ src/main/java/org/openset/automator/test/web/WebContext.java
@@ -1,0 +1,12 @@
+package org.openset.automator.test.web;
+
+import org.openset.automator.settings.web.WebSettings;
+
+public class WebContext {
+    public WebSettings settings;
+    public Browser browser;
+
+    public WebContext(WebSettings settings) {
+
+    }
+}