+++ src/test/java/e2e/ElectronApiDemosTests.java
@@ -1,0 +1,15 @@
+package e2e;
+
+import org.openset.automator.test.electron.ElectronTest;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+public class ElectronApiDemosTests extends ElectronTest {
+
+    @Test
+    @DisplayName("Electron API Demos Smoke Tests")
+    public void smokeTest() {
+        String page = getApp().getDriver().getPageSource();
+        System.out.println(page);
+    }
+}